Create UIMenu.ts
May just use this for all of the ui logic classes.
Most of the things that i do are all routine, unlike working with a gui, the functions that you need are pretty much all the same.

So I think that i will have to create only one shell class for this app